# Small speedups to core.raise_to_shaped().