--- a/jax/core.py
+++ b/jax/core.py
@@ -1331,9 +1331,12 @@ pytype_aval_mappings[Token] = lambda _: abstract_token
 
 
 def raise_to_shaped(aval: AbstractValue, weak_type=None):
+  aval_type = type(aval)
+  if aval_type is ShapedArray and weak_type is None:
+    return aval
   if weak_type is None:
     weak_type = getattr(aval, 'weak_type', False)
-  for typ in type(aval).__mro__:
+  for typ in aval_type.__mro__:
     handler = raise_to_shaped_mappings.get(typ)
     if handler: return handler(aval, weak_type)
   raise TypeError(type(aval))
@@ -1561,7 +1564,7 @@ def canonicalize_shape(shape: Shape, context: str="") -> Shape:
     A tuple of canonical dimension values.
   """
   try:
-    return tuple(map(_canonicalize_dimension, shape))
+    return tuple(unsafe_map(_canonicalize_dimension, shape))
   except TypeError:
     pass
   raise _invalid_shape_error(shape, context)